Modificata la variabile che contrassegna la potenza nominale del file turbine_data.csv da float a int

+++ lib/lettore_csv/libestrazione_csv/include/estrattore_csv.h
@@ -8,7 +8,7 @@
 
     struct turbina {
         char *nome;
-        float potenza_nominale;
+        int potenza_nominale;
         struct turbina *next;
     };
